+++ packages/f-content-cards/src/components/ContentCards.vue
@@ -23,10 +23,12 @@
 
 <script>
 import { globalisationServices } from '@justeat/f-services';
-import initialiseBraze, { logCardClick, logCardImpressions } from '@justeat/f-metadata';
-import ContentCards from '../services/contentCard.service';
+import initialiseBrazeDispatcher from '@justeat/f-metadata';
 import cardTemplates from './cardTemplates';
 import tenantConfigs from '../tenants';
+
+export const CARDSOURCE_BRAZE = 'braze';
+export const CARDSOURCE_CUSTOM = 'custom';
 
 /**
  * Generates card-specific analytics data suitable for sending back to GTM via f-trak
@@ -116,11 +118,16 @@ export default {
             copy: { ...localeConfig },
             titleCard: {},
             hasLoaded: false,
-            loadingCard
+            loadingCard,
+            brazeDispatcher: null
         };
     },
 
     computed: {
+        /**
+         * Determines the tenant based on the currently selected locale in order to choose correct translations
+         * @return {String}
+         **/
         tenant () {
             return {
                 'en-GB': 'uk',
@@ -136,30 +143,66 @@ export default {
     },
 
     watch: {
+        /**
+         * Determines what card impressions should be logged, and whether the loaded flag should be set
+         * @param {Card[]} current
+         * @param {Card[]} previous
+         **/
         cards (current, previous) {
+            this.$emit('get-card-count', current.length);
+
             if (current.length && (current.length !== previous.length)) {
-                logCardImpressions(this.rawCards);
+                this.brazeDispatcher.logCardImpressions(this.cards.map(({ id }) => id));
+            }
+            if ((current.length > 0) && (previous.length === 0)) {
+                this.hasLoaded = true;
+            }
+        },
+
+        /**
+         * Monitors the loaded flag to emit the has-loaded event if necessary
+         * @param {Boolean} current
+         * @param {Boolean} previous
+         **/
+        hasLoaded (current, previous) {
+            if (current && !previous) {
+                this.$emit('has-loaded', true);
             }
         }
     },
 
+    /**
+     * Emits an event that allows consuming code to inject custom content cards, and sets off braze initialisation
+     **/
     mounted () {
+        this.$emit('custom-cards-callback', this.customContentCards.bind(this));
         this.setupBraze(this.apiKey, this.userId);
     },
 
+    /**
+     * Sets up dependencies required by descendant components
+     **/
     provide () {
         const component = this;
 
         return {
+            /**
+             * Reflects card click events though to common click event handler
+             **/
             emitCardClick (card) {
-                component.trackCardClick(card);
-                logCardClick(card);
+                component.handleCardClick(card);
             },
 
+            /**
+             * Reflects card view events though to common view event handler
+             **/
             emitCardView (card) {
-                component.trackCardVisibility(card);
+                component.handleCardView(card);
             },
 
+            /**
+             * Emits voucher code click event with given ongoing url
+             **/
             emitVoucherCodeClick (url) {
                 component.$emit('voucherCodeClick', {
                     url
@@ -171,41 +214,87 @@ export default {
     },
 
     methods: {
+        /**
+         * Initializes braze and handles success / failure states from the returned promise
+         * @param {String} apiKey
+         * @param {String} userId
+         * @param {Boolean} enableLogging
+         * @return {Promise<void>}
+         **/
         setupBraze (apiKey, userId, enableLogging = false) {
-            try {
-                initialiseBraze({
-                    apiKey,
-                    userId,
-                    enableLogging,
-                    callbacks: {
-                        handleContentCards: this.contentCards
-                    }
+            return initialiseBrazeDispatcher({
+                apiKey,
+                userId,
+                enableLogging,
+                enabledCardTypes: this.enabledCardTypes,
+                callbacks: {
+                    handleContentCards: this.brazeContentCards
+                }
+            })
+                .then(dispatcher => {
+                    this.brazeDispatcher = dispatcher;
+                })
+                .catch(error => {
+                    this.$emit('on-error', error);
                 });
-            } catch (error) {
-                this.$emit('on-error', error);
-            }
         },
 
-        contentCards (appboy) {
-            if (!appboy) return;
-            const { cards, rawCards, titleCard } = new ContentCards(appboy, {
-                enabledCardTypes: this.enabledCardTypes
-            })
-                .removeDuplicateContentCards()
-                .filterCards()
-                .getTitleCard()
-                .arrangeCardsByTitles()
-                .applyCardLimit(this.cardLimit)
-                .output();
+        /**
+         * Common method for handling card ingestion to component. Card list length of 0 (after filtering) is considered
+         * 'successful' but does not overwrite any cards currently in place, in order to maintain cards that are present
+         * @param {String} source
+         * @param {Function} successCallback
+         * @param {Function} failCallback
+         * @param {Card[]} cards
+         **/
+        contentCards ({
+            source,
+            successCallback = () => {},
+            failCallback = () => {}
+        }, cards) {
+            if (cards === undefined) {
+                return failCallback();
+            }
 
-            this.rawCards = rawCards;
-            this.cards = cards;
-            this.titleCard = titleCard;
-            this.hasLoaded = true;
+            const limitedCards = this.limitCards(cards, this.cardLimit);
 
-            this.$emit('on-braze-init', appboy);
-            this.$emit('get-card-count', cards.length);
-            this.$emit('has-loaded', true);
+            if ((this.cards.length !== 0) && (limitedCards.length === 0)) {
+                return successCallback();
+            }
+
+            this.cards = limitedCards.map(card => Object.assign(card, { source }));
+
+            return successCallback();
+        },
+
+        /**
+         * Handles card ingestion via braze
+         * @param {Card[]} cards
+         **/
+        brazeContentCards (cards) {
+            this.contentCards({
+                source: CARDSOURCE_BRAZE,
+                successCallback: () => this.$emit('on-braze-init', window.appboy)
+            }, cards);
+        },
+
+        /**
+         * Handles custom card ingestion
+         * @param {Card[]} cards
+         **/
+        customContentCards (cards) {
+            this.contentCards({
+                source: CARDSOURCE_CUSTOM
+            }, cards);
+        },
+
+        /**
+         * Method for returning a shallow copy of the cards array with a maximum of the given limit
+         * @param {Object[]} cards
+         * @param {Number} limit
+         **/
+        limitCards (cards, limit) {
+            return limit > -1 ? cards.slice(0, limit) : cards;
         },
 
         /**
@@ -239,6 +328,45 @@ export default {
             return false;
         },
 
+        /**
+         * Takes appropriate response for click event for given card object based on its source
+         * @param card
+         */
+        handleCardClick (card) {
+            switch (card.source) {
+                case CARDSOURCE_BRAZE:
+                    this.trackBrazeCardClick(card);
+                    this.brazeDispatcher.logCardClick(card.id);
+                    break;
+                case CARDSOURCE_CUSTOM:
+                    // TBC DPDAMI-2688
+                    break;
+                default:
+                    throw new Error('Invalid card source type');
+            }
+        },
+
+        /**
+         * Takes appropriate response for view event for given card object based on its source
+         * @param card
+         */
+        handleCardView (card) {
+            switch (card.source) {
+                case CARDSOURCE_BRAZE:
+                    this.trackBrazeCardVisibility(card);
+                    break;
+                case CARDSOURCE_CUSTOM:
+                    // TBC DPDAMI-2688
+                    break;
+                default:
+                    throw new Error('Invalid card source type');
+            }
+        },
+
+        /**
+         * Uses given pushToDataLayer function to report braze card event
+         * @param payload
+         */
         pushBrazeEvent (payload) {
             this.pushToDataLayer({
                 event: 'BrazeContent',
@@ -248,18 +376,30 @@ export default {
             });
         },
 
-        testIdForItemWithIndex (index) {
-            return this.testId && `ContentCard-${index}`;
-        },
-
-        trackCardClick (card) {
+        /**
+         * Generates a click event for the given card data and reports using the common method
+         * @param card
+         */
+        trackBrazeCardClick (card) {
             const event = createBrazeCardEvent('click', card);
             this.pushBrazeEvent(event);
         },
 
-        trackCardVisibility (card) {
+        /**
+         * Generates a view event for the given card data and reports using the common method
+         * @param card
+         */
+        trackBrazeCardVisibility (card) {
             const event = createBrazeCardEvent('view', card);
             this.pushBrazeEvent(event);
+        },
+
+        /**
+         * Generates a unique test id on a per-card basis if testId prop provided
+         * @param index
+         */
+        testIdForItemWithIndex (index) {
+            return this.testId && `ContentCard-${this.testId}-${index}`;
         }
     }
 };