+++ packages/f-content-cards/src/components/_tests/ContentCards.test.js
@@ -1,7 +1,8 @@
+/* eslint indent: ["error", 4, {ignoredNodes: ["TemplateLiteral > *"]}] */
 import { shallowMount } from '@vue/test-utils';
 import Vue from 'vue';
-import initialiseBraze, { logCardImpressions, logCardClick } from '@justeat/f-metadata';
-import ContentCards from '../ContentCards.vue';
+import initialiseBraze from '@justeat/f-metadata';
+import ContentCards, { CARDSOURCE_BRAZE, CARDSOURCE_CUSTOM } from '../ContentCards.vue';
 import cardTemplates from '../cardTemplates';
 
 jest.mock('@justeat/f-metadata');
@@ -11,27 +12,16 @@ const userId = '__USER_ID__';
 
 const url = '__URL__';
 const button = '__BUTTON__';
-const line1 = '__LINE_1__';
 const description = '__DESCRIPTION__';
 const title = '__TITLE__';
 const voucherCode = '__VOUCHERCODE__';
 const order = '__ORDER__';
 const id = btoa('ABC123');
 
-const createRawCard = type => ({
-    id,
-    url,
-    button,
-    description,
-    title,
-    extras: {
-        order,
-        button_1: button, // eslint-disable-line camelcase
-        custom_card_type: type, // eslint-disable-line camelcase
-        line_1: line1, // eslint-disable-line camelcase
-        voucher_code: voucherCode // eslint-disable-line camelcase
-    }
-});
+const brazeDispatcher = {
+    logCardClick: jest.fn(),
+    logCardImpressions: jest.fn()
+};
 
 const createCard = type => ({
     id,
@@ -44,15 +34,30 @@ const createCard = type => ({
     voucherCode
 });
 
-const createAppboyInstance = cardTypes => ({
-    cards: cardTypes.map(type => createRawCard(type))
-});
+const createBrazeCards = cardTypes => cardTypes.map(type => createCard(type));
 
 beforeEach(() => {
     jest.resetAllMocks();
+    initialiseBraze.mockResolvedValue(brazeDispatcher);
 });
 
 describe('ContentCards', () => {
+    it('should emit a custom content callback when mounted', () => {
+        // Arrange & Act
+        const instance = shallowMount(ContentCards, {
+            propsData: {
+                apiKey,
+                userId
+            }
+        });
+
+        // Assert
+        expect(instance.emitted()['custom-cards-callback']).toBeTruthy(); // Correct event emitted
+        expect(instance.emitted()['custom-cards-callback'].length).toBe(1); // Once
+        expect(instance.emitted()['custom-cards-callback'][0].length).toBe(1); // With one argument
+        expect(instance.emitted()['custom-cards-callback'][0][0]).toBeInstanceOf(Function); // That is a callback
+    });
+
     it('should call intitialiseBraze when mounted', () => {
         // Arrange & Act
         shallowMount(ContentCards, {
@@ -69,10 +74,11 @@ describe('ContentCards', () => {
         expect(settings.enableLogging).toBe(false);
     });
 
-    it('should format and display allowed card types', async () => {
+    it('should handle rejections from the promise returned by intitialiseBraze by emitting an onError event', async () => {
         // Arrange
-        const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
-        const appboy = createAppboyInstance(cardTypes);
+        initialiseBraze.mockReset();
+        const error = new Error('foo');
+        initialiseBraze.mockRejectedValue(error);
 
         // Act
         const instance = shallowMount(ContentCards, {
@@ -81,12 +87,103 @@ describe('ContentCards', () => {
                 userId
             }
         });
-        instance.vm.contentCards(appboy);
+        await instance.vm.$nextTick();
+
+        // Assert
+        expect(instance.emitted()['on-error']).toBeTruthy();
+        expect(instance.emitted()['on-error'].length).toBe(1);
+        expect(instance.emitted()['on-error'][0]).toEqual([error]);
+    });
+
+    it('should format and display allowed card types', async () => {
+        // Arrange
+        const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
+        const cards = createBrazeCards(cardTypes);
+
+        // Act
+        const instance = shallowMount(ContentCards, {
+            propsData: {
+                apiKey,
+                userId
+            }
+        });
+        instance.vm.brazeContentCards(cards);
         await instance.vm.$nextTick();
 
         // Assert
         expect(instance.findAllComponents(cardTemplates.PromotionCard).length).toBe(2);
         expect(instance.findAllComponents(cardTemplates.PostOrderCard).length).toBe(1);
+    });
+
+
+    it('should format and display injected custom cards', async () => {
+        // Arrange
+        const cardTypes = ['Post_Order_Card_1'];
+        const cards = createBrazeCards(cardTypes);
+
+        // Act
+        const instance = shallowMount(ContentCards, {
+            propsData: {
+                apiKey,
+                userId
+            }
+        });
+        instance.vm.customContentCards(cards);
+        await instance.vm.$nextTick();
+
+        // Assert
+        expect(instance.findAllComponents(cardTemplates.PostOrderCard).length).toBe(1);
+    });
+
+    it.each`
+            type                             | component
+            ${'Anniversary_Card_1'}          | ${'VoucherCard'}
+            ${'Voucher_Card_1'}              | ${'VoucherCard'}
+            ${'Restaurant_FTC_Offer_Card'}   | ${'FirstTimeCustomerCard'}
+            ${'Post_Order_Card_1'}           | ${'PostOrderCard'}
+            ${'Promotion_Card_1'}            | ${'PromotionCard'}
+            ${'Promotion_Card_2'}            | ${'PromotionCard'}
+            ${'Terms_And_Conditions_Card'}   | ${'TermsAndConditionsCard'}
+            ${'Terms_And_Conditions_Card_2'} | ${'TermsAndConditionsCard'}
+            ${'Home_Promotion_Card_1'}       | ${'HomePromotionCard1'}
+            ${'Home_Promotion_Card_2'}       | ${'HomePromotionCard2'}
+        `('should map `$type` card type to `$component` component', async ({ type, component }) => {
+        // Arrange
+        const cardTypes = [type];
+        const cards = createBrazeCards(cardTypes);
+
+        // Act
+        const instance = shallowMount(ContentCards, {
+            propsData: {
+                apiKey,
+                userId
+            }
+        });
+        instance.vm.brazeContentCards(cards);
+        await instance.vm.$nextTick();
+
+        // Assert
+        expect(instance.findAllComponents(cardTemplates[component]).length).toBe(1);
+    });
+
+    it('should not render invalid card types as components', async () => {
+        // Arrange
+        const cardTypes = ['INVALID'];
+        const cards = createBrazeCards(cardTypes);
+
+        // Act
+        const instance = shallowMount(ContentCards, {
+            propsData: {
+                apiKey,
+                userId,
+                testId: 'empty-content-cards'
+            }
+        });
+        instance.vm.brazeContentCards(cards);
+        await instance.vm.$nextTick();
+
+        // Assert
+        expect(() => instance.get('[data-test-id="empty-content-cards"] *')).toThrowError();
     });
 
     describe('loading state', () => {
@@ -109,7 +206,7 @@ describe('ContentCards', () => {
         it('should hide the skeleton loading card after Braze has initialised', async () => {
             // Arrange
             const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
-            const appboy = createAppboyInstance(cardTypes);
+            const cards = createBrazeCards(cardTypes);
             const { SkeletonLoader } = cardTemplates;
 
             // Act
@@ -119,7 +216,7 @@ describe('ContentCards', () => {
                     userId
                 }
             });
-            instance.vm.contentCards(appboy);
+            instance.vm.brazeContentCards(cards);
             await instance.vm.$nextTick();
 
             // Assert
@@ -178,7 +275,7 @@ describe('ContentCards', () => {
     it('should call logCardImpressions from f-metadata with data from all displayed cards', async () => {
         // Arrange
         const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
-        const appboy = createAppboyInstance(cardTypes);
+        const cards = createBrazeCards(cardTypes);
 
         // Act
         const instance = shallowMount(ContentCards, {
@@ -187,11 +284,11 @@ describe('ContentCards', () => {
                 userId
             }
         });
-        instance.vm.contentCards(appboy);
+        instance.vm.brazeContentCards(cards);
         await instance.vm.$nextTick();
 
         // Assert
-        expect(logCardImpressions).toHaveBeenCalledWith(appboy.cards);
+        expect(brazeDispatcher.logCardImpressions).toHaveBeenCalledWith(cards.map(card => card.id));
     });
 
     describe('when mounting descendants', () => {
@@ -203,7 +300,7 @@ describe('ContentCards', () => {
                 inject: ['emitCardClick', 'emitCardView', 'emitVoucherCodeClick']
             });
             const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
-            const appboy = createAppboyInstance(cardTypes);
+            const cards = createBrazeCards(cardTypes);
             const instance = shallowMount(ContentCards, {
                 propsData: {
                     apiKey,
@@ -217,7 +314,7 @@ describe('ContentCards', () => {
             const cardClickHandler = jest.spyOn(instance.vm._provided, 'emitCardClick');
             const cardViewHandler = jest.spyOn(instance.vm._provided, 'emitCardView');
             const voucherCodeClickHandler = jest.spyOn(instance.vm._provided, 'emitVoucherCodeClick');
-            instance.vm.contentCards(appboy);
+            instance.vm.brazeContentCards(cards);
             await instance.vm.$nextTick();
 
             return {
@@ -231,6 +328,7 @@ describe('ContentCards', () => {
         it('should provide a view handler', async () => {
             const { instance, cardViewHandler } = await arrange();
             const cardViewDetails = createCard('Post_Order_Card_1');
+            cardViewHandler.mockImplementation(() => {});
 
             // Act
             instance.find('[data-promotion-card="true"]').vm.emitCardView(cardViewDetails);
@@ -239,10 +337,13 @@ describe('ContentCards', () => {
             expect(cardViewHandler).toHaveBeenCalledWith(cardViewDetails);
         });
 
-        describe('the view handler', () => {
+        describe('the view handler, for a braze card', () => {
             it('should push a correctly-formed tracking event to the data layer', async () => {
                 const { instance } = await arrange();
-                const cardViewDetails = createCard('Post_Order_Card_1');
+                const cardViewDetails = {
+                    source: CARDSOURCE_BRAZE,
+                    ...createCard('Post_Order_Card_1')
+                };
 
                 // Act
                 instance.find('[data-promotion-card="true"]').vm.emitCardView(cardViewDetails);
@@ -265,9 +366,44 @@ describe('ContentCards', () => {
             });
         });
 
+        describe('the view handler, for a card with invalid source', () => {
+            it('should throw an error', async () => {
+                const { instance } = await arrange();
+                const cardViewDetails = {
+                    source: 'invalid',
+                    ...createCard('Post_Order_Card_1')
+                };
+
+                // Act
+                expect(() => {
+                    instance.find('[data-promotion-card="true"]').vm.emitCardView(cardViewDetails);
+                })
+                // Assert
+                    .toThrowError();
+            });
+        });
+
+        describe('the view handler, for a card with custom source', () => {
+            it('should not throw an error', async () => {
+                const { instance } = await arrange();
+                const cardViewDetails = {
+                    source: CARDSOURCE_CUSTOM,
+                    ...createCard('Post_Order_Card_1')
+                };
+
+                // Act
+                expect(() => {
+                    instance.find('[data-promotion-card="true"]').vm.emitCardView(cardViewDetails);
+                })
+                // Assert
+                    .not.toThrowError();
+            });
+        });
+
         it('should provide a click handler', async () => {
             const { instance, cardClickHandler } = await arrange();
             const cardClickDetails = createCard('Post_Order_Card_1');
+            cardClickHandler.mockImplementation(() => {});
 
             // Act
             instance.find('[data-promotion-card="true"]').vm.emitCardClick(cardClickDetails);
@@ -276,10 +412,13 @@ describe('ContentCards', () => {
             expect(cardClickHandler).toHaveBeenCalledWith(cardClickDetails);
         });
 
-        describe('the click handler', () => {
+        describe('the click handler, for a braze card', () => {
             it('should push a correctly-formed tracking event to the data layer', async () => {
                 const { instance } = await arrange();
-                const cardClickDetails = createCard('Post_Order_Card_1');
+                const cardClickDetails = {
+                    source: CARDSOURCE_BRAZE,
+                    ...createCard('Post_Order_Card_1')
+                };
 
                 // Act
                 instance.find('[data-promotion-card="true"]').vm.emitCardClick(cardClickDetails);
@@ -303,13 +442,50 @@ describe('ContentCards', () => {
 
             it('should log a card click event with braze', async () => {
                 const { instance } = await arrange();
-                const cardClickDetails = 'foo';
+                const cardClickDetails = {
+                    source: CARDSOURCE_BRAZE,
+                    id: 'foo'
+                };
 
                 // Act
                 instance.find('[data-promotion-card="true"]').vm.emitCardClick(cardClickDetails);
 
                 // Assert
-                expect(logCardClick).toHaveBeenCalledWith('foo');
+                expect(brazeDispatcher.logCardClick).toHaveBeenCalledWith('foo');
+            });
+        });
+
+        describe('the click handler, for a card with invalid source', () => {
+            it('should throw an error', async () => {
+                const { instance } = await arrange();
+                const cardClickDetails = {
+                    source: 'invalid',
+                    ...createCard('Post_Order_Card_1')
+                };
+
+                // Act
+                expect(() => {
+                    instance.find('[data-promotion-card="true"]').vm.emitCardClick(cardClickDetails);
+                })
+                // Assert
+                    .toThrowError();
+            });
+        });
+
+        describe('the click handler, for a card with custom source', () => {
+            it('should not throw an error', async () => {
+                const { instance } = await arrange();
+                const cardClickDetails = {
+                    source: CARDSOURCE_CUSTOM,
+                    ...createCard('Post_Order_Card_1')
+                };
+
+                // Act
+                expect(() => {
+                    instance.find('[data-promotion-card="true"]').vm.emitCardClick(cardClickDetails);
+                })
+                // Assert
+                    .not.toThrowError();
             });
         });
 
@@ -347,7 +523,7 @@ describe('ContentCards', () => {
         beforeAll(async () => {
             // Arrange
             const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
-            const appboy = createAppboyInstance(cardTypes);
+            const cards = createBrazeCards(cardTypes);
 
             // Act
             instance = shallowMount(ContentCards, {
@@ -357,7 +533,7 @@ describe('ContentCards', () => {
                     testId
                 }
             });
-            instance.vm.contentCards(appboy);
+            instance.vm.brazeContentCards(cards);
             await instance.vm.$nextTick();
         });
 
@@ -368,16 +544,16 @@ describe('ContentCards', () => {
 
         it('should generate test id attributes on child content cards components', async () => {
             // Assert
-            expect(instance.find('[data-test-id="ContentCard-0"]').exists()).toBeTruthy();
-            expect(instance.find('[data-test-id="ContentCard-1"]').exists()).toBeTruthy();
-            expect(instance.find('[data-test-id="ContentCard-2"]').exists()).toBeTruthy();
+            expect(instance.find(`[data-test-id="ContentCard-${testId}-0"]`).exists()).toBeTruthy();
+            expect(instance.find(`[data-test-id="ContentCard-${testId}-1"]`).exists()).toBeTruthy();
+            expect(instance.find(`[data-test-id="ContentCard-${testId}-2"]`).exists()).toBeTruthy();
         });
     });
 
     it('should not generate test id attribute on self or children when no test id provided', async () => {
         // Arrange
         const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
-        const appboy = createAppboyInstance(cardTypes);
+        const cards = createBrazeCards(cardTypes);
 
         // Act
         const instance = shallowMount(ContentCards, {
@@ -386,7 +562,7 @@ describe('ContentCards', () => {
                 userId
             }
         });
-        instance.vm.contentCards(appboy);
+        instance.vm.brazeContentCards(cards);
         await instance.vm.$nextTick();
 
         // Assert
@@ -394,17 +570,19 @@ describe('ContentCards', () => {
     });
 
     describe('emitters', () => {
-        const appboy = createAppboyInstance(['Post_Order_Card_1']);
+        const cards = createBrazeCards(['Post_Order_Card_1']);
 
         const testEmitter = async (emitter, expectedArgs) => {
-            // Arrange & Act
+            // Arrange
             const instance = await shallowMount(ContentCards, {
                 propsData: {
                     apiKey,
                     userId
                 }
             });
-            await instance.vm.contentCards(appboy);
+
+            // Act
+            await instance.vm.brazeContentCards(cards);
 
             // Assert
             const [calls] = instance.emitted()[emitter];
@@ -414,7 +592,11 @@ describe('ContentCards', () => {
             expect(args).toEqual(expectedArgs);
         };
 
-        it('should emit an event containing the appboy instance when appboy is initialised', async () => {
+        it('should emit an event containing the global appboy instance when appboy is initialised', async () => {
+            // Arrange
+            const appboy = '__APPBOY__';
+            window.appboy = appboy;
+
             await testEmitter('on-braze-init', appboy);
         });